refactor: move Achievements to components directory
