+++ packages/frontend/app/templates/About/About.jsx
@@ -3,7 +3,8 @@ import Banner from "design-system/components/banner"
 import Button from "design-system/components/button"
 import Navigation from "design-system/blocks/navigation"
 import Footer from "design-system/blocks/footer"
-import Achievements from "./components/AboutUs/Achievements"
+
+import Achievements from "./components/Achievements"
 import AboutUs from "./components/AboutUs"
 import Values from "./components/Values"
 import Supporters from "./components/Supporters"
